+++ src/impls/alloc_impls.rs
@@ -17,6 +17,18 @@ impl_newtype_generic!(T: [], Rc<T>, Rc::new);
 impl_newtype_generic!(T: [], Pin<Box<T>>, Box::pin);
 impl_newtype_generic!(T: [], Pin<Rc<T>>, Rc::pin);
 
+impl<'a, T: Exhaust> Exhaust for alloc::borrow::Cow<'a, T> {
+    type Iter = ::core::iter::Map<<T as Exhaust>::Iter, fn(T) -> alloc::borrow::Cow<'a, T>>;
+
+    /// Note that this implementation necessarily ignores the borrowed versus owned distinction;
+    /// every value returned will be a `Cow::Owned`, not a `Cow::Borrowed`.
+    /// This agrees with the [`PartialEq`] implementation for [`Cow`], which considers
+    /// owned and borrowed to be equal.
+    fn exhaust() -> Self::Iter {
+        <T as Exhaust>::exhaust().map(alloc::borrow::Cow::Owned)
+    }
+}
+
 // Note: This impl is essentially identical to the one for `HashSet`.
 impl<T: Exhaust + Ord> Exhaust for BTreeSet<T> {
     type Iter = ExhaustBTreeSet<T>;